+++ xmtp_mls/src/client.rs
@@ -1,10 +1,5 @@
-use std::{
-    collections::{HashMap, HashSet},
-    mem::Discriminant,
-    pin::Pin,
-};
+use std::{collections::HashSet, mem::Discriminant};
 
-use futures::{Stream, StreamExt};
 use openmls::{
     framing::{MlsMessageIn, MlsMessageInBody},
     group::GroupEpoch,
@@ -19,23 +14,21 @@ use tls_codec::{Deserialize, Error as TlsSerializationError};
 use xmtp_proto::{
     api_client::XmtpMlsClient,
     xmtp::mls::api::v1::{
-        group_message,
         welcome_message::{Version as WelcomeMessageVersion, V1 as WelcomeMessageV1},
         GroupMessage, WelcomeMessage,
     },
 };
 
 use crate::{
-    api_client_wrapper::{ApiClientWrapper, GroupFilter, IdentityUpdate},
+    api_client_wrapper::{ApiClientWrapper, IdentityUpdate},
     groups::{
-        extract_group_id, validated_commit::CommitValidationError, AddressesOrInstallationIds,
-        GroupError, IntentError, MlsGroup, PreconfiguredPolicies,
+        validated_commit::CommitValidationError, AddressesOrInstallationIds, IntentError, MlsGroup,
+        PreconfiguredPolicies,
     },
     identity::Identity,
     storage::{
         db_connection::DbConnection,
         group::{GroupMembershipState, StoredGroup},
-        group_message::StoredGroupMessage,
         refresh_state::EntityKind,
         EncryptedMessageStore, StorageError,
     },
@@ -146,20 +139,6 @@ impl From<String> for ClientError {
 impl From<&str> for ClientError {
     fn from(value: &str) -> Self {
         Self::Generic(value.to_string())
-    }
-}
-
-pub(crate) struct MessagesStreamInfo<'c, ApiClient> {
-    group: MlsGroup<'c, ApiClient>,
-    cursor: u64,
-}
-
-impl<'c, ApiClient> Clone for MessagesStreamInfo<'c, ApiClient> {
-    fn clone(&self) -> Self {
-        Self {
-            group: self.group.clone(),
-            cursor: self.cursor,
-        }
     }
 }
 
@@ -490,128 +469,11 @@ where
             })
             .collect())
     }
-
-    fn process_streamed_welcome(
-        &self,
-        welcome: WelcomeMessage,
-    ) -> Result<MlsGroup<ApiClient>, ClientError> {
-        let welcome_v1 = extract_welcome_message(welcome)?;
-        let conn = self.store.conn()?;
-        let provider = self.mls_provider(&conn);
-
-        MlsGroup::create_from_encrypted_welcome(
-            self,
-            &provider,
-            welcome_v1.hpke_public_key.as_slice(),
-            welcome_v1.data,
-        )
-        .map_err(|e| ClientError::Generic(e.to_string()))
-    }
-
-    pub async fn stream_conversations(
-        &'a self,
-    ) -> Result<Pin<Box<dyn Stream<Item = MlsGroup<ApiClient>> + Send + 'a>>, ClientError> {
-        let installation_key = self.installation_public_key();
-        let id_cursor = 0;
-
-        let subscription = self
-            .api_client
-            .subscribe_welcome_messages(installation_key, Some(id_cursor as u64))
-            .await?;
-
-        let stream = subscription
-            .map(|welcome_result| async {
-                let welcome = welcome_result?;
-                self.process_streamed_welcome(welcome)
-            })
-            .filter_map(|res| async {
-                match res.await {
-                    Ok(group) => Some(group),
-                    Err(err) => {
-                        log::error!("Error processing stream entry: {:?}", err);
-                        None
-                    }
-                }
-            });
-
-        Ok(Box::pin(stream))
-    }
-
-    pub async fn stream_all_messages(
-        &'a self,
-    ) -> Result<Pin<Box<dyn Stream<Item = StoredGroupMessage> + 'a + Send>>, ClientError> {
-        let mut groups_stream = self.stream_conversations().await?;
-        self.sync_welcomes().await?;
-        let mut group_id_to_info: HashMap<Vec<u8>, MessagesStreamInfo<'a, ApiClient>> = self
-            .find_groups(None, None, None, None)?
-            .into_iter()
-            .map(|group| {
-                (
-                    group.group_id.clone(),
-                    MessagesStreamInfo { group, cursor: 0 },
-                )
-            })
-            .collect();
-
-        let filters: Vec<GroupFilter> = group_id_to_info
-            .iter()
-            .map(|(group_id, info)| GroupFilter::new(group_id.clone(), Some(info.cursor)))
-            .collect();
-        let messages_subscription = self.api_client.subscribe_group_messages(filters).await?;
-        let group_id_to_info_clone = group_id_to_info.clone();
-        let stream = messages_subscription
-            .map(move |res| {
-                let group_id_to_info_clone = group_id_to_info_clone.clone();
-                async move {
-                    match res {
-                        Ok(envelope) => {
-                            let group_id = extract_group_id(&envelope)?;
-                            group_id_to_info_clone
-                                .get(&group_id)
-                                .ok_or(ClientError::StreamInconsistency(
-                                    "Received message for a non-subscribed group".to_string(),
-                                ))?
-                                .group
-                                .process_stream_entry(envelope)
-                                .await
-                        }
-                        Err(err) => Err(GroupError::Api(err)),
-                    }
-                }
-            })
-            .filter_map(move |res| async {
-                match res.await {
-                    Ok(Some(message)) => Some(message),
-                    Ok(None) => None,
-                    Err(err) => {
-                        log::error!("Error processing stream entry: {:?}", err);
-                        None
-                    }
-                }
-            });
-
-        Ok(Box::pin(stream))
-        // loop {
-        //     tokio::select! {
-        //         item = groups_stream.next() => {
-        //             match item {
-        //                 Some(convo) => callback.on_conversation(Arc::new(FfiGroup {
-        //                     inner_client: inner_client.clone(),
-        //                     group_id: convo.group_id,
-        //                     created_at_ns: convo.created_at_ns,
-        //                 })),
-        //                 None => break
-        //             }
-        //         }
-        //         _ = &mut close_receiver => {
-        //             break;
-        //         }
-        //     }
-        // }
-    }
 }
 
-fn extract_welcome_message(welcome: WelcomeMessage) -> Result<WelcomeMessageV1, ClientError> {
+pub(crate) fn extract_welcome_message(
+    welcome: WelcomeMessage,
+) -> Result<WelcomeMessageV1, ClientError> {
     match welcome.version {
         Some(WelcomeMessageVersion::V1(welcome)) => Ok(welcome),
         _ => Err(ClientError::Generic(
@@ -646,7 +508,6 @@ fn has_active_installation(updates: &Vec<IdentityUpdate>) -> bool {
 
 #[cfg(test)]
 mod tests {
-    use futures::StreamExt;
     use xmtp_cryptography::utils::generate_local_wallet;
 
     use crate::{
@@ -827,22 +688,5 @@ mod tests {
             bola_messages.get(1).unwrap().decrypted_message_bytes,
             vec![1, 2, 3]
         )
-    }
-
-    #[tokio::test]
-    async fn test_stream_welcomes() {
-        let alice = ClientBuilder::new_test_client(&generate_local_wallet()).await;
-        let bob = ClientBuilder::new_test_client(&generate_local_wallet()).await;
-
-        let alice_bob_group = alice.create_group(None).unwrap();
-
-        let mut bob_stream = bob.stream_conversations().await.unwrap();
-        alice_bob_group
-            .add_members(vec![bob.account_address()])
-            .await
-            .unwrap();
-
-        let bob_received_groups = bob_stream.next().await.unwrap();
-        assert_eq!(bob_received_groups.group_id, alice_bob_group.group_id);
     }
 }
